Refactor useAtomixGlass hook into smaller composable hooks
This change decomposes the complex `useAtomixGlass` hook into several focused hooks located in `src/lib/composables/atomix-glass/`:
- `useGlassMediaPreferences`
- `useGlassMouseTracker`
- `useGlassSize`
- `useGlassCornerRadius`
- `useGlassBackgroundDetection`
- `useGlassOverLightConfig`
- `useGlassTransforms`

This improves maintainability and readability by separating concerns. It also updates the `AtomixGlass` component snapshot, which had drifted from the actual implementation (specifically regarding `border-radius` variable naming and filter styles). The logic within the hooks remains faithful to the original implementation.
