+++ src/lib/composables/atomix-glass/useGlassMediaPreferences.ts
@@ -1,0 +1,84 @@
+import { useEffect, useMemo, useState } from 'react';
+
+interface UseGlassMediaPreferencesProps {
+  reducedMotion: boolean;
+  highContrast: boolean;
+  disableEffects: boolean;
+}
+
+export function useGlassMediaPreferences({
+  reducedMotion,
+  highContrast,
+  disableEffects,
+}: UseGlassMediaPreferencesProps) {
+  const [userPrefersReducedMotion, setUserPrefersReducedMotion] = useState(false);
+  const [userPrefersHighContrast, setUserPrefersHighContrast] = useState(false);
+
+  useEffect(() => {
+    if (typeof window === 'undefined' || typeof window.matchMedia !== 'function') {
+      return undefined;
+    }
+
+    try {
+      const mediaQueryReducedMotion = window.matchMedia('(prefers-reduced-motion: reduce)');
+      const mediaQueryHighContrast = window.matchMedia('(prefers-contrast: high)');
+
+      setUserPrefersReducedMotion(mediaQueryReducedMotion.matches);
+      setUserPrefersHighContrast(mediaQueryHighContrast.matches);
+
+      const handleReducedMotionChange = (e: MediaQueryListEvent) => {
+        setUserPrefersReducedMotion(e.matches);
+      };
+
+      const handleHighContrastChange = (e: MediaQueryListEvent) => {
+        setUserPrefersHighContrast(e.matches);
+      };
+
+      if (mediaQueryReducedMotion.addEventListener) {
+        mediaQueryReducedMotion.addEventListener('change', handleReducedMotionChange);
+        mediaQueryHighContrast.addEventListener('change', handleHighContrastChange);
+      } else if (mediaQueryReducedMotion.addListener) {
+        mediaQueryReducedMotion.addListener(handleReducedMotionChange);
+        mediaQueryHighContrast.addListener(handleHighContrastChange);
+      }
+
+      return () => {
+        try {
+          if (mediaQueryReducedMotion.removeEventListener) {
+            mediaQueryReducedMotion.removeEventListener('change', handleReducedMotionChange);
+            mediaQueryHighContrast.removeEventListener('change', handleHighContrastChange);
+          } else if (mediaQueryReducedMotion.removeListener) {
+            mediaQueryReducedMotion.removeListener(handleReducedMotionChange);
+            mediaQueryHighContrast.removeListener(handleHighContrastChange);
+          }
+        } catch (cleanupError) {
+          console.error('AtomixGlass: Error cleaning up media query listeners:', cleanupError);
+        }
+      };
+    } catch (error) {
+      console.error('AtomixGlass: Error setting up media queries:', error);
+      return undefined;
+    }
+  }, []);
+
+  const effectiveReducedMotion = useMemo(
+    () => reducedMotion || userPrefersReducedMotion,
+    [reducedMotion, userPrefersReducedMotion]
+  );
+
+  const effectiveHighContrast = useMemo(
+    () => highContrast || userPrefersHighContrast,
+    [highContrast, userPrefersHighContrast]
+  );
+
+  const effectiveDisableEffects = useMemo(
+    () => disableEffects || effectiveReducedMotion,
+    [disableEffects, effectiveReducedMotion]
+  );
+
+  return {
+    effectiveReducedMotion,
+    effectiveHighContrast,
+    effectiveDisableEffects,
+  };
+}
